requirements.txt
Django==4.1.6
djangorestframework==3.14.0
mysql-connector-python==8.0.32
mysqlclient==2.1.1
drf-spectacular==0.25.1
requests==2.28.2
pika==1.3.1